add convolve polynomial

+++ pipeline.py
@@ -75,7 +75,7 @@ Even by singling out only the hue you still have a very meaningful representatio
 The end result is a more robust color thresholding over simpler parameters." 
 """
 def color_s_channel(img,s_thresh=(0, 255)):
-    hls = cv2.cvtColor(img, cv2.COLOR_BGR2HLS).astype(np.float)
+    hls = cv2.cvtColor(img, cv2.COLOR_BGR2HLS)
     s_channel = hls[:,:,2]
     s_binary = np.zeros_like(s_channel)
     s_binary[(s_channel >= s_thresh[0]) & (s_channel <= s_thresh[1])] = 1
